Add options to control assert+assertuser parameters
I went back and forth over what should happen when the options are set
to false. In the previous commit, I had made logout() delete assertuser
from the defaultParams, because it’s guaranteed to be wrong now. But by
that logic, login() should also delete assert if the option is false –
if the request passed with an assert parameter, then it must be 'anon'
and therefore wrong now as well. I think it’s better to not touch the
defaultParams if the corresponding option isn’t set – we warn users
against varying the options across calls, but if they do it anyways, any
resulting errors are their faul.

Importing m3api as 'm3api/core.js' will only work out of the box on
Node.js; users wanting to use the library in a browser will need to
configure an import map or use a bundler. I think this is the only
workable option, though (importing e.g. '../m3api/core.js' would be very
fragile, depending on how npm moves modules around in the tree).